Simple Test cases for stage five implemented

+++ CourierLibrary/Model/Order.cs
@@ -7,7 +7,7 @@ namespace CourierLibrary.Model
     public class Order
     {
         public IEnumerable<Parcel> Parcels { get; set; }
-        public decimal Discount { get; set; }
+        public int Discount { get; set; }
         public double TotalCost { get; set; }
         public bool FastSpeed { get; set; }
     }
